TimedReset-luokka toimii. Tällä hetkellä database resetoidaan viiden minuutin välein

+++ RaspberryServer/src/main/java/Ohjelmistoprojekti2/raspberryServer/TimedReset.java
@@ -1,18 +1,28 @@
 package Ohjelmistoprojekti2.raspberryServer;
 
+import javax.annotation.PostConstruct;
+
+import org.springframework.stereotype.Component;
+
 import Ohjelmistoprojekti2.raspberryServer.domain.RaspberryRepository;
 
+@Component
 public class TimedReset {
 	// 3600000 millisekuntia eli tunti
-	final long timeInterval = 10000;
+	final long timeInterval = 300000;
 	RaspberryRepository raspberryRepository;
 
-	// TimedReset reset = new TimedReset() {
-
-	public void run() {
+	//@PostConstruct annotaatiolla saadaan käynnistettyä runReset-metodi beanin käynnistyksen jälkeen
+	@PostConstruct
+	public void runReset() {
 		while (true) {
-			raspberryRepository.deleteAll();
-			System.out.println("Database reset");
+			if(raspberryRepository != null) {
+				raspberryRepository.deleteAll();
+				System.out.println("Database reset");
+			}else {
+				System.out.println("Database already empty");
+			}
+			
 
 			try {
 				Thread.sleep(timeInterval);
@@ -20,6 +30,6 @@ public class TimedReset {
 				e.printStackTrace();
 			}
 		}
-		// }
 	};
+	
 }
